IMplement the internal use case meta
