--- src/app/wizard-creator/components/use-case-form-components/interal-dev/wizard-use-case-internal-dev-form.component.html
@@ -1,8 +1,0 @@
-<div class="dynamic-coding-use-case-outer-container">
-  <ng-container [ngSwitch]="userCase">
-    <app-wizard-use-case-internal-custom-component
-      *ngSwitchCase="wizardUseCase.UseCaseCustomComponent"
-    ></app-wizard-use-case-internal-custom-component>
-
-  </ng-container>
-</div>
