+++ src/app/wizard-creator/components/wizard-creator-form-dynamic-elements/wizard-creator-form-dynamic-elements.component.html
@@ -3,13 +3,5 @@
     <app-wizard-use-case-general-writing-form
      *ngSwitchCase="useCaseGroupEnum.GeneralWriting"
     ></app-wizard-use-case-general-writing-form>
-
-    <app-wizard-use-case-internal-dev-form
-       *ngSwitchCase="useCaseGroupEnum.InternalDev"
-    >
-
-    </app-wizard-use-case-internal-dev-form>
-   
-
   </ng-container>
 </div>
