+++ src/app/wizard-creator/wizard-creator.module.ts
@@ -22,8 +22,6 @@ import { WizardCreatorSelectorFieldComponent } from './components/common/wizard-
 import { WizardUseCaseService } from './services/use-case/wizard-use-case.service';
 import { WizardFormService } from './services/wizard-form.service';
 import { UseCaseFormBaseComponent } from './components/use-case-form-components/use-case-form-base/use-case-form-base.component';
-import { WizardUseCaseInternalCustomComponentComponent } from './components/use-case-form-components/interal-dev/children/wizard-use-case-internal-custom-component.component';
-import { WizardUseCaseInternalDevFormComponent } from './components/use-case-form-components/interal-dev/wizard-use-case-internal-dev-form.component';
 
 
 
@@ -42,8 +40,6 @@ import { WizardUseCaseInternalDevFormComponent } from './components/use-case-for
     WizardUseCaseGeneralWritingFormMessageComponent,
     WizardCreatorSelectorFieldComponent,
     UseCaseFormBaseComponent,
-    WizardUseCaseInternalCustomComponentComponent,
-    WizardUseCaseInternalDevFormComponent,
   ],
   imports: [
     CommonModule,
